Widen explicitly in the FFT butterfly, and in two spectrum locals
Thirteen -Wdouble-promotion warnings across the Linux, macOS and headless
jobs, in code I added one commit after documenting that the tree was
warning-free and that "warning-free" is a claim about the compiler in front
of you. MSVC does not emit this one, so the local build was clean.

The arithmetic is unchanged. The four butterfly inputs are now widened once,
visibly, instead of promoting eight times implicitly where each float
element meets a double twiddle -- which reads better as well as compiling
quieter. The FFT still agrees with the naive DFT bin for bin.

+++ app/src/SpectrumWidget.cpp
@@ -118,7 +118,7 @@ void SpectrumWidget::paintEvent(QPaintEvent* event) {
     painter.setPen(Qt::NoPen);
 
     for (std::size_t band = 0; band < bands.size(); ++band) {
-        const double level = bands[band];
+        const double level = static_cast<double>(bands[band]);
         if (level <= 0.0) {
             continue;
         }
@@ -131,7 +131,7 @@ void SpectrumWidget::paintEvent(QPaintEvent* event) {
     // its own colour; the same idea, in a colour that reads on the gradient.
     painter.setPen(QColor(255, 255, 255, 190));
     for (std::size_t band = 0; band < peaks.size(); ++band) {
-        const double peak = peaks[band];
+        const double peak = static_cast<double>(peaks[band]);
         if (peak <= 0.0) {
             continue;
         }
